refactor as createBackendXXX
