+++ packages/backend-indexer/src/index.ts
@@ -1,1 +1,4 @@
-export {};
+export * from "./createIndexerBackend.js";
+export * from "./isPublicRpcMethod.js";
+
+export * from "./controllers/index.js";
