+++ packages/backend-wallet/README.md
@@ -1,2 +1,5 @@
-# Bundler
-Owl Protocol ERC4337 Bundler
+# Wallet Backend
+
+[Wallet Client](https://viem.sh/docs/clients/wallet.html) designed to work with [viem](https://viem.sh/) or as a simple [EIP1193](https://eips.ethereum.org/EIPS/eip-1193) provider.
+
+Created by adding multiple [Account](https://viem.sh/docs/accounts/local) instances which enables it to support multiple accounts.
