+++ packages/backend-wallet/src/createMultiAccountWalletClient.ts
@@ -23,31 +23,6 @@ import {
 import { omit } from "lodash-es";
 import { transactionRequestDecodeZod } from "@owlprotocol/zod-sol";
 
-export type WalletRpcMethod = (typeof walletRpcMethods)[number];
-
-export const walletRpcMethods = [
-    "eth_accounts",
-    "eth_requestAccounts",
-    "eth_sendTransaction",
-    "eth_sign",
-    "eth_signTransaction",
-    "eth_signTypedData_v4",
-    "personal_sign",
-    "wallet_addEthereumChain",
-    "wallet_getPermissions",
-    "wallet_switchEthereumChain",
-    "wallet_watchAsset",
-] as const;
-
-/**
- * Check if RPC method is for wallets.
- * @param method
- * @returns true if wallet rpc method
- */
-export function isWalletRpcMethod(method: string): method is WalletRpcMethod {
-    return walletRpcMethods.includes(method as any);
-}
-
 export type MultiAccountConfig = ClientConfig<Transport, Chain, Account> & {
     /** Public client override. Use this instead of instantiating a new one */
     publicClient?: PublicClient<Transport, Chain>;
