+++ packages/backend-wallet/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./createMultiAccountWalletClient.js";
+export * from "./isWalletRpcMethod.js";
